Rename files & cleanup
