--- src/components/chat/ChatListItem.js
@@ -1,72 +1,0 @@
-import * as React from "react";
-import Box from "@mui/joy/Box";
-import ListDivider from "@mui/joy/ListDivider";
-import ListItem from "@mui/joy/ListItem";
-import ListItemButton, { ListItemButtonProps } from "@mui/joy/ListItemButton";
-import Stack from "@mui/joy/Stack";
-import Typography from "@mui/joy/Typography";
-import CircleIcon from "@mui/icons-material/Circle";
-import AvatarWithStatus from "./AvatarWithStatus";
-import { toggleMessagesPane } from "../../utils";
-
-export default function ChatListItem(props) {
-  const { id, sender, messages, selectedChatId, setSelectedChat } = props;
-  const selected = selectedChatId === id;
-  return (
-    <React.Fragment>
-      <ListItem>
-        <ListItemButton
-          onClick={() => {
-            toggleMessagesPane();
-            setSelectedChat({ id, sender, messages });
-          }}
-          selected={selected}
-          color="neutral"
-          sx={{
-            flexDirection: "column",
-            alignItems: "initial",
-            gap: 1,
-          }}
-        >
-          <Stack direction="row" spacing={1.5}>
-            <AvatarWithStatus online={sender.online} src={sender.avatar} />
-            <Box sx={{ flex: 1 }}>
-              <Typography level="title-sm">{sender.name}</Typography>
-              <Typography level="body-sm">{sender.username}</Typography>
-            </Box>
-            <Box
-              sx={{
-                lineHeight: 1.5,
-                textAlign: "right",
-              }}
-            >
-              {messages[0]?.unread && (
-                <CircleIcon sx={{ fontSize: 12 }} color="primary" />
-              )}
-              <Typography
-                level="body-xs"
-                display={{ xs: "none", md: "block" }}
-                noWrap
-              >
-                5 mins ago
-              </Typography>
-            </Box>
-          </Stack>
-          <Typography
-            level="body-sm"
-            sx={{
-              display: "-webkit-box",
-              WebkitLineClamp: "2",
-              WebkitBoxOrient: "vertical",
-              overflow: "hidden",
-              textOverflow: "ellipsis",
-            }}
-          >
-            {messages[0]?.content}
-          </Typography>
-        </ListItemButton>
-      </ListItem>
-      <ListDivider sx={{ margin: 0 }} />
-    </React.Fragment>
-  );
-}
